test_project/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django_extensions==3.2.3
